feat: Good kitten Photo.

+++ code/include/bound.hpp
@@ -33,7 +33,7 @@ public:
     }
     
     // 还没进来就已出去则无交点
-    bool intersect(const Ray &r, double &t_min) {
+    bool intersect(const Ray &r, float &t_min) {
         Vector3f o(r.getOrigin()), invDir(1 / r.getDirection());
         Vector3f t0 = (bounds[0] - o) * invDir;
         Vector3f t1 = (bounds[1] - o) * invDir;
