+++ code/include/curve.hpp
@@ -42,7 +42,7 @@ public:
         }
     }
 
-    bool intersect(const Ray& r, Hit& h, double tmin) override
+    bool intersect(const Ray& r, Hit& h, float tmin) override
     {
         return false;
     }
