+++ code/include/group.hpp
@@ -26,7 +26,7 @@ public:
     ~Group() override {}
 
     // KD-Tree + 包围盒加速求最近交点
-    bool intersect(const Ray &r, Hit &h, double tmin) override {
+    bool intersect(const Ray &r, Hit &h, float tmin) override {
         return kdTree->intersect(r, h, TMIN);
     }
