+++ code/include/transform.hpp
@@ -31,7 +31,7 @@ public:
     ~Transform() {
     }
 
-    virtual bool intersect(const Ray &r, Hit &h, double tmin) {
+    virtual bool intersect(const Ray &r, Hit &h, float tmin) {
         // 1. 将仿射变换逆变换作用到射线上，得到一根新的光线
         Vector3f trSource = transformPoint(transform, r.getOrigin());
         Vector3f trDirection = transformDirection(transform, r.getDirection());
